requirements.txt
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
PyJWT==2.3.0

freezegun~=1.1.0